moves ShareholderSelector to modal directory
