+++ components/ManageBufferModule.tsx
@@ -7,7 +7,7 @@ import {
 } from "@/app/payload-builder/payloadHelperFunctions";
 import { JsonViewerEditor } from "@/components/JsonViewerEditor";
 import { TokenSelector } from "@/components/TokenSelector";
-import { ShareholderSelector } from "@/components/ShareholderSelector";
+import { ShareholderSelector } from "@/components/modal/ShareholderSelector";
 import { NETWORK_OPTIONS, networks } from "@/constants/constants";
 import { getNetworksWithCategory } from "@/lib/data/maxis/addressBook";
 import {
